Move the Token service to the services folder
